+++ dev/workspace/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @prosopo/workspace
 
+## 3.1.19
+### Patch Changes
+
+- 7d5eb3f: bump
+
 ## 3.1.18
 ### Patch Changes