+++ integration/frameworks/vue/vue-procaptcha-wrapper/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @prosopo/vue-procaptcha-wrapper
 
+## 1.1.43
+### Patch Changes
+
+  - @prosopo/procaptcha-wrapper@2.6.45
+
+## 1.1.42
+### Patch Changes
+
+  - @prosopo/procaptcha-wrapper@2.6.44
+
+## 1.1.41
+### Patch Changes
+
+- e01227b: add turbo
+  - @prosopo/procaptcha-wrapper@2.6.43
+
+## 1.1.40
+### Patch Changes
+
+- 7d5eb3f: bump
+- Updated dependencies [7d5eb3f]
+  - @prosopo/procaptcha-wrapper@2.6.42
+
 ## 1.1.39
 ### Patch Changes