+++ packages/detector/CHANGELOG.md
@@ -1,5 +1,49 @@
 # @prosopo/detector
 
+## 3.3.24
+### Patch Changes
+
+- c6faa77: Fix
+
+## 3.3.23
+### Patch Changes
+
+- caf53e2: Async user creation
+
+## 3.3.22
+### Patch Changes
+
+- ea5f1f8: Fix detectors
+
+## 3.3.21
+### Patch Changes
+
+- Updated dependencies [141e462]
+  - @prosopo/types@3.7.2
+
+## 3.3.20
+### Patch Changes
+
+- Updated dependencies [345b25b]
+  - @prosopo/types@3.7.1
+
+## 3.3.19
+### Patch Changes
+
+- 5ad6f48: Fix detect webview iphones
+- e01227b: add turbo
+- Updated dependencies [ce70a2b]
+- Updated dependencies [c2b940f]
+- Updated dependencies [f6b5094]
+  - @prosopo/types@3.7.0
+
+## 3.3.18
+### Patch Changes
+
+- 7d5eb3f: bump
+- Updated dependencies [7d5eb3f]
+  - @prosopo/types@3.6.4
+
 ## 3.3.17
 ### Patch Changes