+++ packages/redis-client/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @prosopo/redis-client
 
+## 1.0.12
+### Patch Changes
+
+  - @prosopo/common@3.1.27
+
+## 1.0.11
+### Patch Changes
+
+- 7d5eb3f: bump
+- Updated dependencies [7d5eb3f]
+  - @prosopo/common@3.1.26
+
 ## 1.0.10
 ### Patch Changes